klu/requirements.txt
asgiref==3.7.2
Django==4.2.5
mysql-connector-python==8.1.0
protobuf==4.21.12
sqlparse==0.4.4
tzdata==2023.3
